Add files via upload
This is mainly an example for using the Raspberry Pi Pico microcontroller with an attached SSD1306 controlled Oled-Display.
As hardware only a Pi Pico and an Oled display with 128 x 64 pixels are needed, the software is based on the Pico SDK, anything else is part of the example.

The code renders stages of a cellular automaton known as "Conway's Game of Life" using pixels as "cells". Starting from a random pattern you will see establishing stable and moving / "living" structures based on very simple rules. 

I named it "Urknall" (big bang) since everything in nature (atoms, crystals, living cells, stars) need just some simple rules to grow to astonishing very complex and often expressive and beautiful stable structures - like the ornaments for this game.